Created a modal for uploading files.
The old tool threw up a file open dialog but didn't give
the user any context as to what was happening. With the new modal dialog
componenet it made sense to try and help the user a bit more.

Changes:
1. Moved the upload tool out of tools.jsx and into its
   own module.  This cleans up the organization as there is
   a dependent Modal class now in upload.jsx.
2. Small tweak to the modal CSS that fixes an embedding issue.
   Now it's possible to put the dialog div nearly anywhere in the DOM
   and it will render correctly.
