+++ src/gm3/components/catalog/tools.jsx
@@ -31,116 +31,6 @@ import * as mapActions from '../../actions/map';
 
 import * as util from '../../util';
 
-/** Upload features to a vector layer from a file
- *  on the user's hard drive.
- *
- *  Currently supports KML and GeoJSON.
- *
- */
-export class UploadTool extends Component {
-
-    constructor() {
-        super();
-        this.startUpload = this.startUpload.bind(this); 
-    }
-
-    parseFiles() {
-        // grab the list of files.
-        const files = this.refs.fileInput.files;
-
-        // setup a file reader.
-        const reader = new FileReader();
-
-        // ugly closure bridge, but it let me recycle some
-        //  code.
-        const self = this;
-
-        // the current implementation only uses a "single" file
-        //  upload, but these leaves the door open to a multiple
-        //  file functionality.
-        for(let i = 0, f; f = files[i]; i++) {
-            // TODO: This should probably be changed to use
-            //   () => {} style functions to prevent the context change,
-            //   instead of abusing "self".
-            // setup a "load" event.
-            reader.onload = (function(file) {
-                return function(e) {
-                    // geojson is always used as the output format.
-                    const geojson_format = new ol.format.GeoJSON();
-
-                    // input_format is defaulted to null and only
-                    //  set if the file format can be inferred.
-                    let input_format = null;
-                    if(file.type.indexOf('kml') >= 0) {
-                        // sweet, KML file.
-                        input_format = new ol.format.KML();
-                    } else {
-                        // try to see if it is JSON parseable
-                        try {
-                            JSON.parse(e.target.result);
-                            input_format = new ol.format.GeoJSON();
-                        } catch(err) {
-                            // swallow the exception.
-                        }
-                    }
-
-                    if(input_format !== null) {
-                        // parse the features.
-                        let ol_features = input_format.readFeatures(e.target.result);
-
-                        // TODO: This should be getting the map projection
-                        //       from the map-view!!!
-                        for(let f of ol_features) {
-                            f.setGeometry(f.getGeometry().transform('EPSG:4326', 'EPSG:3857'));
-                        }
-                        // internal feature representation is as GeoJSON, so the parsed
-                        //  ol feautures need converted here...
-                        let collection = geojson_format.writeFeaturesObject(ol_features);
-
-                        // only add features to the first layer listed in the
-                        //  catalog's layer definition
-                        let src = self.props.layer.src[0];
-                        // kick off the event
-                        self.props.store.dispatch(msActions.addFeatures(
-                            src.mapSourceName, src.layerName, collection.features));
-
-                        // TODO: Notify the user that n-number of features 
-                        //       were added to the layer.
-                        // this.store.dispatch(createNotice('info', '...'));
-                    } else {
-                        // TODO: Notify the user that something has
-                        //       gone terribly wrong.
-                        // this.store.dispatch(createNotice('error', '...'));
-                    }
-                }
-            })(f);
-
-            // trigger the actual file read.
-            reader.readAsText(f);
-        }
-    }
-
-    startUpload() {
-        // open the upload dialog.
-        this.refs.fileInput.click();
-
-        // setup the response function that will read
-        //   the upload and add it to the layer.
-        this.refs.fileInput.onchange = (evt) => {
-            this.parseFiles();
-        };
-    }
-
-    render() {
-        let tip = 'Add features from file';
-        return (
-            <i className='upload tool' onClick={this.startUpload} title={tip}>
-                <input ref='fileInput' className='hide' type='file'/>
-            </i>
-        );
-    }
-}
-
 /** Generic class for basic "click this, do this" tools.
  */
 class Tool extends Component {
