+++ src/less/modal.less
@@ -1,8 +1,12 @@
 
 .modal-blocker {
-    position: absolute;
+    position: fixed;
+    z-index: 1000;
+
     top: 0; left: 0;
     width: 100%; height: 100%;
+    min-height: 100%;
+
     background-color: rgba(0,0,0,.75);
     text-align: center;
 }
